Repaired faulty code in 'doStepLoop function'

+++ src/main.cpp
@@ -4,9 +4,10 @@ void loop()
 {
     if (ESP_BT.available())
     {
-        float incoming = ESP_BT.read(); // Read byte from cell-phone
+        int incoming = ESP_BT.read(); // Read byte from cell-phone     
         processIncoming(incoming);
     }
 
-    doStepLoop();
+   doStepLoop();
+
 }
